Removed reg wizard and moved the state up one level with the new state management. The license panel is now hooked up.
