+++ UI/Web/src/app/admin/kavita-plus/license/license.component.html
@@ -2,10 +2,11 @@
 
   @let hasLicenseOnFile = licenseService.hasLicenseOnFile();
 
-  @if(!hasLicenseOnFile) {
-    <app-registration-wizard />
-  }
-  @else {
+  @if (!hasLicenseOnFile && activeStep() === KavitaPlusRegistrationStep.Upsell) {
+    <app-kavita-plus-upsell (stepChanged)="activeStep.set($event)" />
+  } @else if (activeStep() === KavitaPlusRegistrationStep.ConnectProviders) {
+    <app-kavita-plus-connect-providers (done)="activeStep.set(KavitaPlusRegistrationStep.Upsell)" />
+  } @else {
 
     <app-license-dashboard />
