+++ UI/Web/src/app/admin/kavita-plus/register-license-key/register-license-key.component.html
@@ -12,7 +12,7 @@
       </div>
     }
 
-    <app-edit-license-key [hasLicense]="hasLicenseOnFile" (updated)="updateFormData($event)" />
+    <app-edit-license-key (updated)="updateFormData($event)" />
 
     <button class="btn btn-primary float-end mt-3" [disabled]="!formIsValid() || isLoading()" (click)="save()">
       @if (isLoading()) {
