--- UI/Web/src/app/admin/kavita-plus/registration-wizard/registration-wizard.component.html
@@ -1,8 +1,0 @@
-@switch (activeStep()) {
-  @case (KavitaPlusRegistrationStep.Upsell) {
-    <app-kavita-plus-upsell (stepChanged)="activeStep.set($event)" />
-  }
-  @case (KavitaPlusRegistrationStep.ConnectProviders) {
-    <app-kavita-plus-connect-providers />
-  }
-}
